# frontend/src/types/openapi.json
{
  "components": {
    "schemas": {
...
      "CveDeltaResponse": {
        "description": "Response from GET /cve-delta \u2014 used by TideWatch integration.",
        "properties": {
          "cutoff_time": {
            "title": "Cutoff Time",
            "type": "string"
          },
          "scans": {
            "items": {
              "$ref": "#/components/schemas/CveDeltaScanEntry"
            },
            "title": "Scans",
            "type": "array"
          },
          "since_hours": {
            "title": "Since Hours",
            "type": "integer"
          },
          "summary": {
            "$ref": "#/components/schemas/CveDeltaSummary"
          },
          "total_scans": {
            "title": "Total Scans",
            "type": "integer"
          }
        },
        "required": [
          "since_hours",
          "cutoff_time",
          "total_scans",
          "summary",
          "scans"
        ],
        "title": "CveDeltaResponse",
        "type": "object"
      },
      "CveDeltaScanEntry": {
        "description": "Individual scan entry in a CVE delta response.",
        "properties": {
          "container_name": {
            "title": "Container Name",
            "type": "string"
          },
          "cves_fixed": {
            "items": {
              "type": "string"
            },
            "title": "Cves Fixed",
            "type": "array"
          },
          "cves_fixed_count": {
            "title": "Cves Fixed Count",
            "type": "integer"
          },
          "cves_introduced": {
            "items": {
              "type": "string"
            },
            "title": "Cves Introduced",
            "type": "array"
          },
          "cves_introduced_count": {
            "title": "Cves Introduced Count",
            "type": "integer"
          },
          "image": {
            "title": "Image",
            "type": "string"
          },
          "scan_date": {
            "title": "Scan Date",
            "type": "string"
          },
          "scan_id": {
            "title": "Scan Id",
            "type": "integer"
          },
          "total_vulns": {
            "title": "Total Vulns",
            "type": "integer"
          }
        },
        "required": [
          "scan_id",
          "scan_date",
          "container_name",
          "image",
          "total_vulns",
          "cves_fixed",
          "cves_fixed_count",
          "cves_introduced",
          "cves_introduced_count"
        ],
        "title": "CveDeltaScanEntry",
        "type": "object"
      },
      "CveDeltaSummary": {
        "description": "Summary totals for CVE delta response.",
        "properties": {
          "net_change": {
            "title": "Net Change",
            "type": "integer"
          },
          "total_cves_fixed": {
            "title": "Total Cves Fixed",
            "type": "integer"
          },
          "total_cves_introduced": {
            "title": "Total Cves Introduced",
            "type": "integer"
          }
        },
        "required": [
          "total_cves_fixed",
          "total_cves_introduced",
          "net_change"
        ],
        "title": "CveDeltaSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Scan": {
        "description": "Full scan schema.",
        "properties": {
          "container_id": {
            "title": "Container Id",
            "type": "integer"
          },
          "critical_count": {
            "title": "Critical Count",
            "type": "integer"
          },
          "cves_fixed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cves Fixed"
          },
          "cves_introduced": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cves Introduced"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "fixable_vulns": {
            "title": "Fixable Vulns",
            "type": "integer"
          },
          "high_count": {
            "title": "High Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_scanned": {
            "title": "Image Scanned",
            "type": "string"
          },
          "low_count": {
            "title": "Low Count",
            "type": "integer"
          },
          "medium_count": {
            "title": "Medium Count",
            "type": "integer"
          },
          "scan_date": {
            "format": "date-time",
            "title": "Scan Date",
            "type": "string"
          },
          "scan_duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scan Duration Seconds"
          },
          "scan_status": {
            "title": "Scan Status",
            "type": "string"
          },
          "total_vulns": {
            "title": "Total Vulns",
            "type": "integer"
          }
        },
        "required": [
          "container_id",
          "image_scanned",
          "id",
          "scan_date",
          "scan_status",
          "scan_duration_seconds",
          "error_message",
          "total_vulns",
          "fixable_vulns",
          "critical_count",
          "high_count",
          "medium_count",
          "low_count"
        ],
        "title": "Scan",
        "type": "object"
      },
      "ScanAbortResponse": {
        "description": "Response from POST /{scan_id}/abort.",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "scan_id": {
            "title": "Scan Id",
            "type": "integer"
          }
        },
        "required": [
          "message",
          "scan_id"
        ],
        "title": "ScanAbortResponse",
        "type": "object"
      },
      "ScanJobSchema": {
        "description": "Schema for scan job status responses.",
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "container_id": {
            "title": "Container Id",
            "type": "integer"
          },
          "container_name": {
            "title": "Container Name",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "scan_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scan Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "container_id",
          "container_name",
          "status",
          "created_at"
        ],
        "title": "ScanJobSchema",
        "type": "object"
      },
...
      "ScanRequest": {
        "description": "Schema for requesting a scan.",
        "properties": {
          "container_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Container Ids"
          }
        },
        "title": "ScanRequest",
        "type": "object"
      },
      "ScanRetryResponse": {
        "description": "Response from POST /{scan_id}/retry.",
        "properties": {
          "container": {
            "title": "Container",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "scan_id": {
            "title": "Scan Id",
            "type": "integer"
          }
        },
        "required": [
          "message",
          "scan_id",
          "container"
        ],
        "title": "ScanRetryResponse",
        "type": "object"
      },
...
      "ScanTrendsResponse": {
        "description": "Response from GET /scans/trends.",
        "properties": {
          "series": {
            "items": {
              "$ref": "#/components/schemas/ScanTrendDataPoint"
            },
            "title": "Series",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/ScanTrendSummary"
          },
          "velocity": {
            "$ref": "#/components/schemas/ScanTrendVelocity"
          },
          "window_days": {
            "title": "Window Days",
            "type": "integer"
          }
        },
        "required": [
          "window_days",
          "series",
          "summary",
          "velocity"
        ],
        "title": "ScanTrendsResponse",
        "type": "object"
      },
      "ScanTriggerResponse": {
        "description": "Response from POST /scan \u2014 scan enqueue result.",
        "properties": {
          "job_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Job Ids",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "queued": {
            "title": "Queued",
            "type": "integer"
          },
          "skipped": {
            "title": "Skipped",
            "type": "integer"
          },
          "total_requested": {
            "title": "Total Requested",
            "type": "integer"
          }
        },
        "required": [
          "message",
          "queued",
          "skipped",
          "total_requested",
          "job_ids"
        ],
        "title": "ScanTriggerResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Docker vulnerability scanner and remediation dashboard powered by Trivy",
    "title": "VulnForge",
    "version": "4.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scans/cve-delta": {
      "get": {
        "description": "Get CVE delta information from recent scans.\n\nReturns a summary of CVEs fixed and introduced across all containers\nwithin the specified time window. Used by TideWatch for integration.\n\nArgs:\n    since_hours: Number of hours to look back (default 24, max 720/30 days)\n    container_name: Optional filter for specific container\n    scan_id: Optional filter for a specific scan (deterministic retrieval)\n\nReturns:\n    List of scan deltas with container info, cves_fixed, and cves_introduced",
        "operationId": "get_cve_delta_api_v1_scans_cve_delta_get",
        "parameters": [
          {
            "description": "Hours to look back for scan deltas",
            "in": "query",
            "name": "since_hours",
            "required": false,
            "schema": {
              "default": 24,
              "description": "Hours to look back for scan deltas",
              "maximum": 720,
              "minimum": 1,
              "title": "Since Hours",
              "type": "integer"
            }
          },
          {
            "description": "Filter by container name",
            "in": "query",
            "name": "container_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by container name",
              "title": "Container Name"
            }
          },
          {
            "description": "Filter by specific scan ID",
            "in": "query",
            "name": "scan_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by specific scan ID",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CveDeltaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cve Delta",
        "tags": [
          "Scans"
        ]
      }
    },
    "/api/v1/scans/history/{container_id}": {
      "get": {
        "description": "Get scan history for a container.",
        "operationId": "get_scan_history_api_v1_scans_history__container_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "container_id",
            "required": true,
            "schema": {
              "title": "Container Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Scan"
                  },
                  "title": "Response Get Scan History Api V1 Scans History  Container Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scan History",
        "tags": [
          "Scans"
        ]
      }
    },
    "/api/v1/scans/jobs/{job_id}": {
      "get": {
        "description": "Get status of a scan job by its correlation ID.\n\nExternal consumers (TideWatch) poll this endpoint to track scan progress\nand retrieve the linked scan_id once processing completes.\n\nRate limit: 120 requests per minute (read endpoint, consistent with /current).",
        "operationId": "get_scan_job_status_api_v1_scans_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanJobSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scan Job Status",
        "tags": [
          "Scans"
        ]
      }
    },
...
    "/api/v1/scans/scan": {
      "post": {
        "description": "Trigger a scan of containers using the scan queue.\n\nRate limit: 10 requests per minute to prevent scan spam.\n\nDelegates to ScanOrchestrator which handles ScanJob creation, batch\nregistration, and enqueuing in a single two-phase commit.",
        "operationId": "scan_containers_api_v1_scans_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanTriggerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scan Containers",
        "tags": [
          "Scans"
        ]
      }
    },
...
    "/api/v1/scans/trends": {
      "get": {
        "description": "Return aggregated scan trends for dashboards.",
        "operationId": "get_scan_trends_api_v1_scans_trends_get",
        "parameters": [
          {
            "in": "query",
            "name": "window_days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 90,
              "minimum": 1,
              "title": "Window Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanTrendsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scan Trends",
        "tags": [
          "Scans"
        ]
      }
    },
    "/api/v1/scans/{scan_id}/abort": {
      "post": {
        "description": "Abort a running or queued scan.\n\nArgs:\n    scan_id: ID of the scan to abort\n\nReturns:\n    Success message if scan was aborted",
        "operationId": "abort_scan_api_v1_scans__scan_id__abort_post",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanAbortResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Abort Scan",
        "tags": [
          "Scans"
        ]
      }
    },
    "/api/v1/scans/{scan_id}/retry": {
      "post": {
        "description": "Retry a failed scan.\n\nArgs:\n    scan_id: ID of the scan to retry\n\nReturns:\n    Success message if scan was queued for retry",
        "operationId": "retry_scan_api_v1_scans__scan_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRetryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Scan",
        "tags": [
          "Scans"
        ]
      }
    },
...
  }
}